nyc-taxi-project/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.3
pandas==2.3.3
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
Flask==2.3.2
python-dateutil==2.8.2
pandas==2.2.2
